# openapi.yaml
openapi: 3.1.0
info:
  title: auth
  version: "1.1.0"
  description: >-
    User accounts, sessions and JWT-based authentication for the atc-web platform. Callers are
    trusted application backends holding an API key (never an end user's browser directly); they
    forward the end user's address and agent via `X-Client-IP` / `X-Client-User-Agent` for the
    audit log when the presented key carries the `proxy` flag. Access tokens (ES256 JWTs) and
    opaque refresh tokens are both returned in the JSON response body — this service never sets
    cookies.
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/users:
    post:
      operationId: auth.users.create
      summary: Register a user
      description: >-
        Requires an API key with the `write` role (`write` or `readwrite`; `read`-only keys get
        403 FORBIDDEN). Registration succeeds even when the verification email fails to send —
        `verificationEmailSent` tells the caller whether to offer a resend.
      tags: [Users]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreateRequest" }
      responses:
        "201":
          description: User created.
          headers:
            location:
              schema: { type: string, example: "/v1/users/8c9e2b7e-..." }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateUserResponse" }
        "400":
          description: VALIDATION_FAILED (malformed body) or WEAK_PASSWORD (policy rejected).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: EMAIL_TAKEN — an account with this email already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
              example: { error: { code: EMAIL_TAKEN, message: "an account with this email already exists" } }
        "429": { $ref: "#/components/responses/RateLimited" }

    get:
      operationId: auth.users.list
      summary: List or look up users
      description: >-
        Any valid API key may call this — there is no role gate on reads (confirmed: a
        `read`-only key can enumerate every user, including by exact-email lookup). Pass `email`
        for an exact-match lookup (returns 0 or 1 items, `nextCursor` always `null`); otherwise
        paginates by keyset cursor, newest first.
      tags: [Users]
      x-stack-visibility: authenticated
      parameters:
        - name: email
          in: query
          schema: { $ref: "#/components/schemas/Email" }
          description: Exact-match lookup; mutually exclusive in effect with pagination.
        - name: limit
          in: query
          schema: { type: string, pattern: "^([1-9]|[1-9][0-9]|100)$" }
          description: Page size, 1-100. Default 20.
        - name: cursor
          in: query
          schema: { type: string, maxLength: 128 }
          description: Opaque keyset cursor from a previous page's `nextCursor`.
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserListResponse" }
        "400":
          description: VALIDATION_FAILED (malformed query) or INVALID_CURSOR (undecodable cursor).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
              examples:
                validation:
                  value: { error: { code: VALIDATION_FAILED, message: "querystring/limit must match pattern \"^([1-9]|[1-9][0-9]|100)$\"", details: [] } }
                invalidCursor:
                  summary: "INVALID_CURSOR is a plain thrown object, not an AuthError instance — no `details` field (architectural inconsistency vs. every other domain error in this service, confirmed in source)."
                  value: { error: { code: INVALID_CURSOR, message: "invalid cursor" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      operationId: auth.users.get
      summary: Get a user by id
      tags: [Users]
      x-stack-visibility: authenticated
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/UserNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }

    patch:
      operationId: auth.users.patch
      summary: Update a user's name or status
      description: >-
        Requires an API key with the `write` role. Setting `status: "disabled"` revokes every
        active session for the user as part of the same transaction.
      tags: [Users]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserPatchRequest" }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/UserNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }

    delete:
      operationId: auth.users.delete
      summary: Delete a user
      description: Requires an API key with the `write` role.
      tags: [Users]
      x-stack-visibility: authenticated
      responses:
        "204": { description: User deleted. }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/UserNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/users/{id}/sessions:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      operationId: auth.sessions.list
      summary: List a user's active sessions
      description: Sessions that are neither revoked nor expired, as of now.
      tags: [Sessions]
      x-stack-visibility: authenticated
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionListResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/UserNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }

    delete:
      operationId: auth.sessions.revokeAll
      summary: Revoke every active session for a user
      description: >-
        Returns the count revoked. Note this endpoint returns `200` with a body, **not** `204` —
        unlike the single-session revoke below, the handler sends a JSON object (confirmed in
        source: no `reply.code(204)` on this route). Requires an API key with the `write` role.
      tags: [Sessions]
      x-stack-visibility: authenticated
      responses:
        "200":
          description: OK — sessions revoked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RevokeAllResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/UserNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/users/{id}/sessions/{sid}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - name: sid
        in: path
        required: true
        schema: { type: string, format: uuid }
        description: Session id.
    delete:
      operationId: auth.sessions.revokeOne
      summary: Revoke a single session
      description: >-
        Idempotent create-then-delete semantics aside, a session that does not exist or does not
        belong to `id` returns `SESSION_NOT_FOUND` — this route does not separately check whether
        `id` itself refers to a real user first. Requires an API key with the `write` role.
      tags: [Sessions]
      x-stack-visibility: authenticated
      responses:
        "204": { description: Session revoked (or already revoked). }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: SESSION_NOT_FOUND — no such session for this user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
              example: { error: { code: SESSION_NOT_FOUND, message: "session not found" } }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/users/{id}/events:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      operationId: auth.events.list
      summary: List a user's audit events
      description: >-
        Keyset-paginated, newest first. Note the pagination field is `nextBefore`, not
        `nextCursor` — this service is one of several across the platform whose list envelopes use
        a slightly different cursor field name (a known, not-yet-normalized cross-service
        inconsistency).
      tags: [Events]
      x-stack-visibility: authenticated
      parameters:
        - name: limit
          in: query
          schema: { type: string, pattern: "^([1-9]|[1-9][0-9]|100)$" }
          description: Page size, 1-100. Default 50.
        - name: before
          in: query
          schema: { type: string, pattern: "^[0-9]{1,15}$" }
          description: Event id cursor from a previous page's `nextBefore`.
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventListResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/UserNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/login:
    post:
      operationId: auth.login
      summary: Log in with email and password
      description: >-
        Any valid API key may call this (no role gate — login is an end user's own action, not a
        back-office mutation). On success, issues a new session and returns both tokens in the
        body.
      tags: [Authentication]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401":
          description: >-
            UNAUTHORIZED (missing/invalid API key) or INVALID_CREDENTIALS (email or password is
            incorrect — the same message and status whether the email is unknown or the password
            is wrong, and the password hasher runs a dummy hash for an unknown email so response
            time does not reveal which).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "403":
          description: >-
            ACCOUNT_DISABLED, or EMAIL_NOT_VERIFIED (only when `LOGIN_REQUIRES_VERIFIED_EMAIL=true`,
            default `false`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "423":
          description: >-
            ACCOUNT_LOCKED — too many failed attempts (`LOGIN_MAX_FAILURES`, default 10). Carries
            `retry-after` and `details.retryAfterSec`.
          headers:
            retry-after:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
              example: { error: { code: ACCOUNT_LOCKED, message: "too many failed attempts, try again later", details: { retryAfterSec: 900 } } }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/refresh:
    post:
      operationId: auth.refresh
      summary: Rotate a refresh token
      description: >-
        A refresh token that has already been rotated (replayed) revokes the whole session
        (`TOKEN_REUSED`) — this is a reuse-detection signal, not a benign retry.
      tags: [Authentication]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: OK — a new access/refresh token pair for the same session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401":
          description: >-
            UNAUTHORIZED (API key), INVALID_TOKEN (unknown, revoked, or expired refresh token), or
            TOKEN_REUSED (an already-rotated token was replayed; the session is revoked as a
            side effect).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "403":
          description: ACCOUNT_DISABLED — the owning account was disabled since the session was created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
              example: { error: { code: ACCOUNT_DISABLED, message: "account is disabled" } }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/logout:
    post:
      operationId: auth.logout
      summary: Revoke the session owning a refresh token
      description: Idempotent — an unknown or already-revoked refresh token still returns 204.
      tags: [Authentication]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "204": { description: Session revoked, or was already revoked/unknown. }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/introspect:
    post:
      operationId: auth.introspect
      summary: Validate an access token
      description: >-
        Never throws a domain error — an invalid, expired, or otherwise unusable token is reported
        in the 200 body as `active: false`, not as an HTTP error status. See
        `IntrospectResponse` for the exact, source-verified field combinations.
      tags: [Authentication]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntrospectRequest" }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntrospectResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/verify-email:
    post:
      operationId: auth.emailVerification.verify
      summary: Consume an email-verification token
      tags: [Email verification]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenRequest" }
      responses:
        "200":
          description: OK — the user, now verified.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/InvalidToken" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/verify-email/resend:
    post:
      operationId: auth.emailVerification.resend
      summary: Re-send the verification email
      description: >-
        Silent (still `202`) when the email is unknown, to avoid account enumeration. A known,
        already-verified email returns `409 ALREADY_VERIFIED` instead — this is itself a (minor,
        two-way) enumeration signal, confirmed and flagged in the Phase 0 audit, not fixed here.
        A per-user cooldown (`RESEND_COOLDOWN_SEC`, default 60s) throttles repeat requests with
        `429 TOO_MANY_REQUESTS`; a separate, unrelated self-throttle (`429 RATE_LIMITED`) can also
        apply per API key.
      tags: [Email verification]
      x-stack-visibility: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailOnlyRequest" }
      responses:
        "202":
          description: >-
            Accepted — either the mail was queued, or the email is unknown (both look identical to
            the caller by design).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AcceptedResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: ALREADY_VERIFIED — this email belongs to an already-verified account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
              example: { error: { code: ALREADY_VERIFIED, message: "email is already verified" } }
        "429": { $ref: "#/components/responses/ResendThrottled" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <API key>`. Keys are `id:secret[:role[:proxy]]`; `role` is
        `read`/`write`/`readwrite` (missing role defaults to `readwrite`). Only
        `POST /v1/users`, `PATCH /v1/users/{id}` and `DELETE /v1/users/{id}` require the `write`
        role (or `readwrite`) — every other `/v1/*` route and `/metrics` accept any valid key
        regardless of role; this sub-scoping is noted per-operation since a single Bearer scheme
        cannot express it structurally.
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string, description: "Stable machine-readable error code." }
            message: { type: string }
            details:
              description: >-
                Present only for some errors. An object for a single domain error's structured
                detail (e.g. `{retryAfterSec}`, `{problems}`), or an array of `{path, message,
                params}` for `VALIDATION_FAILED`. `INVALID_CURSOR` never has `details` (it is a
                plain thrown object, not a domain `AuthError` instance).
              oneOf:
                - type: object
                - type: array
                  items: { type: object }
...
    UserCreateRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { $ref: "#/components/schemas/Email" }
        password: { type: string, minLength: 1, maxLength: 1024, description: "Also checked against the password policy (minimum length, denylist, not the account's own email, ...)." }
        name: { type: [string, "null"], minLength: 1, maxLength: 120 }
...
    CreateUserResponse:
      type: object
      additionalProperties: false
      required: [user, verificationEmailSent]
      properties:
        user: { $ref: "#/components/schemas/User" }
        verificationEmailSent:
          type: boolean
          description: False when the mail transport rejected the send; registration itself still succeeded.
...